fix: make data_set_basics playbook idempotent
The tasks that created data sets or files before being copied to were causing the copy tasks to fail, as the destination already existed.

It suggested adding 'force: true', but that caused other problems.

These changes make it so that the playbook doesn't have these errors, and makes it fully idempotent for re-running smoothly.

Signed-off-by: Jacob Emery <jacob.emery@ibm.com>