fix expiration date for locally signed certs
Realized that the expiration date was being set to 1 year out instead of what was specified by new_expiry_date when signed by local CA. GENCERT command needed the expiration date as well.

Also removed handlers for setropts refresh so that it could be done before a listing of the renewed certificate at the end to verify visually in the output that renewal was successful.

Signed-off-by: Jacob Emery <jacob.emery@ibm.com>